try move constant to other file
